requirements.txt
facebook-business>=17.0.0
python-dotenv>=1.0.0
jsonschema>=4.0.0
Pillow>=10.0.0
